rewrite with a more regex-based approach
The previous approach was not robust enough for SMILES with more
than one pair of square brackets.  The current rewritten approach
-- though constrained to C, N, O, S -- no longer carries this
limitation and is more careful (even anticipating) for elements
enclosed in square brackets and usually written with two characters
(e.g., Sn).

As checked against the test data and a `diff -s`, the output of
SMILES about saturated compounds is the same as with the previous
version of the script.